Refactor: 不要なレンダリングを避けるためにaudio関連のコンポーネントに切り分ける
